Moves all integration tests into one file
